[WIP] general infrastructure to use Repr
The idea is to have the informations about the [Repr.t] in the [Drv.t].
That is the code for declaring the [Repr.t] (i.e. `Repr.int`) and names
for the [Repr.t], equality and comparison functions.

The [Repr.t] is unstaged and the equality and comparison functions are
declared at the top of the generated module.
We can then call the correponding function when translating terms.

- add a [repr] field to the driver along with some functions
- a [repr] can be parametrized ([Forall] constructor)
- [repr] for types in stdlib (not all, waiting for Sets and Bags) has been added
- TODO: add [repr] for types used in the module (defined types and arguments)
- TODO: handle specialisation of the parametrized [repr]
- TODO: generate only equality and comparison that are used